Erro de arredondamento int
No arquivo "eclipse_nv1", linha 260, a variável raioPlanetaPixel não pode ser inteiro, isso resulta em uma diferença muito grande na profundidade do trânsito.

+++ scripts/func.c
@@ -55,7 +55,7 @@ double curvaLuz(double x0, double y0, int tamanhoMatriz, int raioPlanetaPixel, d
 	return valor;
 }
 
-double curvaLuzLua(double x0, double y0, double xm, double ym, double rMoon, int tamanhoMatriz, int raioPlanetaPixel, double *estrelaManchada, double *kk, double maxCurvaLuz){
+double curvaLuzLua(double x0, double y0, double xm, double ym, double rMoon, int tamanhoMatriz, double raioPlanetaPixel, double *estrelaManchada, double *kk, double maxCurvaLuz){
 	double valor = 0;
 	int i;
